rename smartcard pkg to globalplatform
